+++ src/commands/wipe.ts
@@ -7,7 +7,6 @@ import { MissingWorkspaceError } from '../helpers'
 
 export default async (args): Promise<void> => {
   let workspace: Workspace
-  let client: RestClient
 
   try {
     workspace = await Workspace.init(
@@ -21,16 +20,16 @@ export default async (args): Promise<void> => {
     return MissingWorkspaceError(args.workspace)
   }
 
-  client = new RestClient(workspace.config, workspace.name)
+  const client = new RestClient(workspace.config, workspace.name)
 
-  let spinner = ora({
+  const spinner = ora({
     prefixText: `Wiping...`,
     text: `reading files...`,
   }).start()
 
   let files: FileResource[]
   try {
-    files = await client.getAllFiles()
+    files = await client.getAllFiles({ fields: 'path' })
 
     if (workspace.config.enablePaths && workspace.config.enablePaths.length > 0) {
       files = files.filter((file: FileResource): boolean =>
@@ -40,16 +39,15 @@ export default async (args): Promise<void> => {
 
     if (workspace.config.skipPaths && workspace.config.skipPaths.length > 0) {
       files = files.filter(
-        (file: FileResource): boolean =>
-          !workspace.config.skipPaths.some((path): boolean => file.path.startsWith(path)),
+        (file: FileResource): boolean => !workspace.config.skipPaths.some((path) => file.path?.startsWith(path)),
       )
     }
 
-    for (let file of files) {
-      if (workspace.config.ignoreSpecs && file.path.startsWith('specs')) {
+    for (const file of files) {
+      if (workspace.config.ignoreSpecs && file.path?.startsWith('specs')) {
         continue
       }
-      spinner.text = file.path
+      spinner.text = file.path || file.id || 'file'
       await client.deleteFile(file)
     }
   } catch (e) {