+++ src/core/HTTP/RestClient.ts
@@ -1,9 +1,10 @@
 import { IWorkspaceConfig } from '../WorkspaceConfig'
-import { IRestResponse } from './RestInterfaces'
+import { IGetAllFilesParams, IRestResponse } from './RestInterfaces'
 import FileResource, { FileResourceJSON } from './Resources/FileResource'
-import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from 'axios'
+import axios, { AxiosRequestConfig, AxiosResponse, AxiosError, AxiosInstance } from 'axios'
 import { Agent as HTTPAgent } from 'http'
 import { Agent as HTTPSAgent } from 'https'
+import { MAX_CONTENT_LENGTH_MB, ONE_MB } from '../constants'
 
 export class RestClientError<T> extends Error {
   public response: IRestResponse<T>
@@ -15,7 +16,7 @@ export class RestClientError<T> extends Error {
 }
 
 export default class RestClient {
-  public client
+  public client: AxiosInstance
   public clientHeaders
   public clientUrl: string
   public workspaceName: string
@@ -31,7 +32,7 @@ export default class RestClient {
       console.log(
         'upstream is deprecated and will cease to function in a later release. Please use kong_admin_url (upstream url without the workspace suffix)',
       )
-      let match = workspaceConfig.upstream.match(/.*\//)
+      const match = workspaceConfig.upstream.match(/.*\//)
       if (!match) {
         console.log('unable to parse upstream url')
         throw new Error()
@@ -63,7 +64,8 @@ export default class RestClient {
       baseURL: this.clientUrl,
       headers: this.clientHeaders,
       httpAgent: new HTTPAgent({ keepAlive: true }),
-      maxContentLength: 10 * 1000000, // 10mb is kong file system
+      maxContentLength:
+        (workspaceConfig.maxContentLengthInMb ? workspaceConfig.maxContentLengthInMb : MAX_CONTENT_LENGTH_MB) * ONE_MB,
       httpsAgent,
     })
   }
@@ -72,11 +74,13 @@ export default class RestClient {
     return this.handleResponse(await this.client.get(`${this.workspaceName}/files`, options))
   }
 
-  public async getAllFiles<T>(): Promise<FileResourceJSON[]> {
-    let res = await this.client.get(`${this.workspaceName}/files`)
+  public async getAllFiles(params: IGetAllFilesParams = {}): Promise<FileResourceJSON[]> {
+    let res = await this.client.get(`${this.workspaceName}/files`, {
+      params,
+    })
     let files: FileResourceJSON[] = this.handleResponse(res)
     while (res.data.next) {
-      // url already has workspace
+      // url already has workspace and query params
       res = await this.client.get(res.data.next)
       files = files.concat(this.handleResponse(res))
     }
@@ -85,11 +89,23 @@ export default class RestClient {
   }
 
   public async saveFile<Output>(file: FileResource, options: AxiosRequestConfig = {}): Promise<void> {
-    await this.client.put(`${this.workspaceName}/files/${file.path}`, file, options)
+    try {
+      await this.client.put(`${this.workspaceName}/files/${file.path}`, file, options)
+    } catch (e) {
+      console.log(`\n\nError uploading file: ${file.path}\n`)
+      console.error((e.response && e.response.data) || e)
+      throw e
+    }
   }
 
   public async deleteFile<T>(file: FileResource, options: AxiosRequestConfig = {}): Promise<void> {
-    await this.client.delete(`${this.workspaceName}/files/${file.path}`, options)
+    try {
+      await this.client.delete(`${this.workspaceName}/files/${file.path}`, options)
+    } catch (e) {
+      console.log(`\n\nError deleting file: ${file.path}\n`)
+      console.error((e.response && e.response.data) || e)
+      throw e
+    }
   }
 
   public async enablePortal(options: AxiosRequestConfig = {}): Promise<void> {