+++ src/core/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_CONTENT_LENGTH_MB = 10 // 10mb (default) is kong file system
+export const ONE_MB = 1000000