+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.1] - 2023-03-30
+### Fixed
+- Fix issue with `sync` command `--ignore-specs` flag where all remaining files were ignored after ignoring specs
+
+## [3.6.0] - 2023-01-23
+### Added
+- Add new command `sync` to compare local and remote files when deploying
+
+### Changed
+- Update `wipe` command to use the `fields` query param if available
+
+## [3.5.0] - 2022-11-15
+### Added
+- Add conf variable `max_content_length_mb` for adjusting the allowed content length in request bodies
+
+## [3.4.0] - 2022-07-11
+### Added
+- Add env variable `KONG_ADMIN_TOKEN_FILE` for loading Kong Admin Token from file path
+- Log errors from response when uploading or deleting remote files
+
 ## [3.3.1] - 2021-12-03
 ### Fixed
 - Use HTTP keep-alive to improve performance